Moved upgrade handler to its own directory.
